start of updating layout of the schedule and styling it
