+++ src/components/all.sass
@@ -145,6 +145,19 @@ ul
     p
       text-align: center
 
+.card-schedule-item
+  .time-wrapper
+    margin-right: .75rem !important
+    min-width: 8ch
+
+    p:first-child time
+      font-weight: bold
+
+  .schedule-item-wrapper
+    border-left: 2px solid $pyohio-red
+    border-radius: 5px
+    padding-left: 10px
+
 // talk page fixes
 article.speaker-bio
   align-items: center
@@ -230,9 +243,8 @@ article.speaker-bio
 .supporter-name
   text-align: center
   margin: .5rem
-  background-color:  $pyohio-red
+  background-color: $pyohio-red
   padding: .5rem 1rem
   border-radius: 3px
   color: #fff
   font-size: 1.5rem
-
