+++ src/templates/schedule-page.js
@@ -1,0 +1,64 @@
+import React from "react"
+import { graphql } from "gatsby"
+import Helmet from "react-helmet"
+import Layout from "../components/Layout"
+import ScheduleSlot from "../components/ScheduleSlot"
+
+export default class SlottedSchedule extends React.Component {
+  render() {
+    const { data } = this.props
+    const { edges: slots } = data.allSlots
+    const pageTitle = "PyOhio 2019 Schedule"
+
+    return (
+      <Layout>
+        <Helmet>
+          <title>{pageTitle}</title>
+          <meta name="twitter:title" content={pageTitle} />
+          <meta property="og:title" content={pageTitle} />
+        </Helmet>
+        <section className="section">
+          <div className="container">
+            <div className="content">
+              <h1 className="has-text-weight-bold is-size-2">{pageTitle}</h1>
+            </div>
+            {slots.map(({ node: slot }) => (
+              <ScheduleSlot slot={slot} key={slot.id}/>
+            ))}
+          </div>
+        </section>
+      </Layout>
+    )
+  }
+}
+
+export const pageQuery = graphql`
+  query SlotsQuery {
+    allSlots(
+      filter: { kind: { ne: "nothing" } }
+      sort: { fields: [start, room_order] }
+    ) {
+      edges {
+        node {
+          id
+          title
+          kind
+          end
+          duration
+          description_html
+          presentation_id
+          room
+          rooms
+          section
+          speaker_name
+          speakers {
+            name
+            speaker_id
+          }
+          start
+          cancelled
+        }
+      }
+    }
+  }
+`
